lots of optical candy lots of base stuff

+++ install_pure.sh
@@ -23,18 +23,19 @@ then
 	exit 1
 fi
 
-echo -e "\n${cyan}Checking for apps needed by install: \c"
-if [ ! -n "$( which lsb_release 2>/dev/null )" ];
+echo -e "\n${cyan}Checking needed apps: \c"
+if [ -z "$( which lsb_release 2>/dev/null )" ];
 then
-	if [ ! -n "$( which aptitude 2>/dev/null )" ];
+	if [ -n "$( which aptitude 2>/dev/null )" ];
 	then
-		aptitude install lsb-release -P
-	elif [ ! -n "$( which emerge 2>/dev/null )" ];
+		aptitude install lsb-release -P || exit 1
+	elif [ -n "$( which emerge 2>/dev/null )" ];
 	then
-		emerge lsb-release -av
+		emerge lsb-release -av || exit 1
+	else
+		echo -e "\n${RED}Unable to install lsb_release${NC}"
+		exit 1
 	fi
-	echo -e "\t${RED}Please install lsb_release${NC}"
-	exit 1
 else
 	echo -e "${CYAN}Done${NC}"
 
@@ -49,7 +50,7 @@ else
 #			LSBID="redhat"
 #		;;
 		*)
-			echo -e "${RED}Unsupported distribution: $LSBID${NC}"
+			echo -e "${RED}Unsupported distribution: $LSBID${NC}; or lsb_release not found."
 			exit 1
 		;;
 	esac
@@ -57,15 +58,18 @@ else
 	echo -e "${cyan}Found supported distribution family: ${CYAN}$LSBID${NC}"
 fi
 
-if [ ! -n "$( which uudecode 2>/dev/null )" ]; then
+if [ -z "$( which uudecode 2>/dev/null )" ]; then
 	if [ "$LSBID" == "debian" ];
 	then
-		echo -e "\t${RED}Installing uudecode (aptitude install sharutils)${NC}"
-		aptitude install sharutils -P || exit 0
+		echo -e "${RED}Installing uudecode (aptitude install sharutils)${NC}"
+		aptitude install sharutils -P || exit 1
 	elif [ "$LSBID" == "gentoo" ];
 	then
-		echo -e "\t${RED}Installing uudecode (sharutils)${NC}"
+		echo -e "${RED}Installing uudecode (sharutils)${NC}"
 		emerge sharutils -av || exit 1
+	else
+		echo -e "\n${RED}Unable to install uuencode${NC}"
+		exit 1
 	fi
 fi
 
@@ -83,37 +87,45 @@ echo -e "${cyan}Extracting files: \c"
 	tail -n +$payload_start $0 | uudecode | tar -C $TARGETDIR/install -xz || exit 0
 echo -e "${CYAN}Done${NC}"
 
-echo -e "${cyan}Putting files in place: \c${NC}"
+echo -e "${cyan}Putting files in place${NC}\c"
+function movevar {
+	oldvar=$(egrep "$2" $TARGETDIR/$1 | sed 's/\&/\\\&/g')
+	newvar=$(egrep "$2" $TARGETDIR/$1.new | sed 's/\&/\\\&/g')
+	if [  -n "$oldvar" ]; then
+		sed -e "s|$newvar|$oldvar|g" $TARGETDIR/$1.new > $TARGETDIR/$1.tmp
+		mv $TARGETDIR/$1.tmp $TARGETDIR/$1.new
+		echo -e "  ${cyan}$1:  ${CYAN}$oldvar${NC}"
+	fi
+}
+
 if [ -e $TARGETDIR/setup.sh ]; then
-	echo -e "\n  Comparing the old and the new config:"
+	echo -e "\n${cyan}Checking old configuration"
 	mv $TARGETDIR/install/setup.sh $TARGETDIR/setup.sh.new
 
-	echo -e "    ${cyan}Keeping vars:${NC}" # ADMINMAIL,BACKUPDIR,DEBUG,SCRIPTSDIR,MOUNTO,UMOUNTO"
-
-	function movevar {
-		oldvar=$(egrep "$1" $TARGETDIR/setup.sh | sed 's/\&/\\\&/g')
-		newvar=$(egrep "$1" $TARGETDIR/setup.sh.new | sed 's/\&/\\\&/g')
-		if [  -n "$oldvar" ]; then
-			sed -e "s|$newvar|$oldvar|g" $TARGETDIR/setup.sh.new > $TARGETDIR/setup.sh.tmp
-			mv $TARGETDIR/setup.sh.tmp $TARGETDIR/setup.sh.new
-			echo -e "      ${blue}$oldvar${NC}"
-		fi
-	}
-
-	movevar '^export ADMINMAIL=.*$'
-	movevar '^export BACKUPDIR=.*$'
-	movevar '^export DEBUG=.*$'
-	movevar '^export SCRIPTSDIR=.*$'
-	movevar '^export OXIMIRROR=.*$'
-	movevar '^\s*MOUNTO=.*$'
-	movevar '^\s*UMOUNTO=.*$'
+	movevar "setup.sh" '^export ADMINMAIL=.*$'
+	movevar "setup.sh" '^export BACKUPDIR=.*$'
+	movevar "setup.sh" '^export DEBUG=.*$'
+	movevar "setup.sh" '^export SCRIPTSDIR=.*$'
+	movevar "setup.sh" '^export OXIMIRROR=.*$'
+	movevar "setup.sh" '^export OXICOLOR=.*$'
 
 	mv $TARGETDIR/setup.sh.new $TARGETDIR/setup.sh
 else
 	mv $TARGETDIR/install/setup.sh $TARGETDIR/setup.sh
 fi
 
-mv $TARGETDIR/install/backup.sh $TARGETDIR/backup.sh
+if [ -e $TARGETDIR/backup.sh ]; then
+	mv $TARGETDIR/install/backup.sh $TARGETDIR/backup.sh.new
+
+	movevar "backup.sh" '^\s*MOUNTO=.*$'
+	movevar "backup.sh" '^\s*UMOUNTO=.*$'
+
+	mv $TARGETDIR/backup.sh.new $TARGETDIR/backup.sh
+else
+	mv $TARGETDIR/install/backup.sh $TARGETDIR/backup.sh
+fi
+
+#mv $TARGETDIR/install/backup.sh $TARGETDIR/backup.sh
 mv $TARGETDIR/install/init.sh $TARGETDIR/init.sh
 mv $TARGETDIR/install/virtualbox.sh $TARGETDIR/virtualbox.sh
 
@@ -126,14 +138,14 @@ rmdir $TARGETDIR/install/gentoo
 mv $TARGETDIR/install/user/* $TARGETDIR/user
 rmdir $TARGETDIR/install/user
 
-echo -e "\n${cyan}In case of an update, handle old jobfiles${NC}"
+echo -e "\n${cyan}Checking old jobfiles${NC}"
 for FILEPATH in $(ls $TARGETDIR/install/jobs/*.sh); do
 FILE=$(basename $FILEPATH)
 	if [ -e $TARGETDIR/jobs/$FILE ]; then
 		if [ ! -n "$(diff -q $TARGETDIR/jobs/$FILE $TARGETDIR/install/jobs/$FILE)" ]; then
 			mv $TARGETDIR/install/jobs/$FILE $TARGETDIR/jobs/$FILE
 		else
-			echo -e "${RED}->${NC}\t\t${red}$FILE is edited${NC}"
+			echo -e "${RED}->${NC}    ${red}$FILE is edited${NC}"
 			mv $TARGETDIR/install/jobs/$FILE $TARGETDIR/jobs/$FILE.new
 		fi
 	else
@@ -153,8 +165,8 @@ echo -e "\n${cyan}Setting rights: \c"
 	chmod 644 $TARGETDIR/virtualbox.sh
 	chmod 644 $TARGETDIR/setup.sh
 	chmod -R 750 $TARGETDIR/jobs/
-	chmod -R 750 $TARGETDIR/debian/
-	chmod -R 750 $TARGETDIR/gentoo/
+	chmod -R 755 $TARGETDIR/debian/
+	chmod -R 755 $TARGETDIR/gentoo/
 	chmod -R 755 $TARGETDIR/user/
 
 	chown -R root.root $TARGETDIR
@@ -165,45 +177,45 @@ if [ "$LSBID" == "debian" ];
 then
 	if [ ! -e /etc/init.d/oxivbox ];
 	then
-		echo -e "\t${cyan}Activating debian vbox job${NC}"
+		echo -e "  ${cyan}Activating debian vbox job${NC}"
 		ln -s $TARGETDIR/debian/oxivbox.sh /etc/init.d/oxivbox
 	fi
 
-	echo -e "\t${cyan}Activating weekly update check: \c"
+	echo -e "  ${cyan}Activating weekly update check: \c"
 	ln -sf $TARGETDIR/debian/updatecheck.sh /etc/cron.weekly/updatecheck
 	echo -e "${CYAN}Done${NC}"
 
 	if [ -e /var/cache/apt/archives/ ]; then
-		echo -e "\t${cyan}Activating weekly cleanup of /var/cache/apt/archives/: \c"
+		echo -e "  ${cyan}Activating weekly cleanup of /var/cache/apt/archives/: \c"
 		ln -sf $TARGETDIR/debian/cleanup-apt.sh /etc/cron.weekly/cleanup-apt
 		echo -e "${CYAN}Done${NC}"
 	fi
 fi
 ##monthly cron$
-echo -e "\t${cyan}Activating monthly backup statistic: \c"
+echo -e "  ${cyan}Activating monthly backup statistic: \c"
     ln -sf $TARGETDIR/jobs/backup-info.sh /etc/cron.monthly/backup-info
 echo -e "${CYAN}Done${NC}"
 
 ##weelky cron
-echo -e "\t${cyan}Activating weekly backup cleanup (saves a lot of space!): \c"
+echo -e "  ${cyan}Activating weekly backup cleanup (saves a lot of space!): \c"
     ln -sf $TARGETDIR/jobs/backup-cleanup.sh /etc/cron.weekly/backup-cleanup
 echo -e "${CYAN}Done${NC}"
 
 #daily cron
-echo -e "\t${cyan}Activating daily system, ~/scripts and ~/bin backup: \c"
+echo -e "  ${cyan}Activating daily system, ~/scripts and ~/bin backup: \c"
     ln -sf $TARGETDIR/jobs/backup-system.sh /etc/cron.daily/backup-system
     ln -sf $TARGETDIR/jobs/backup-scripts.sh /etc/cron.daily/backup-scripts
 echo -e "${CYAN}Done${NC}"
 
 
 if [ $(which ejabberdctl 2>/dev/null ) ]; then
-    echo -e "\t${CYAN}Found ejabberd, installing daily backup and weekly avatar cleanup${NC}"
+    echo -e "  ${CYAN}Found ejabberd, installing daily backup and weekly avatar cleanup${NC}"
     ln -sf $TARGETDIR/jobs/cleanup-avatars.sh /etc/cron.weekly/cleanup-avatars
     ln -sf $TARGETDIR/jobs/backup-ejabberd.sh /etc/cron.daily/backup-ejabberd
 fi
 
 if [ $(which masqld 2>/dev/null ) ]; then
-    echo -e "\t${CYAN}Found mysql, installing daily backup${NC}"
+    echo -e "  ${CYAN}Found mysql, installing daily backup${NC}"
     ln -sf $TARGETDIR/jobs/backup-mysql.sh /etc/cron.daily/backup-mysql
 fi
 
@@ -214,14 +226,14 @@ echo -e "\n${cyan}Finding all .bashrc files to add init.sh:${NC}"
 function addtorc {
 	if [ ! -n "$(grep oxiscripts/init.sh $1)" ];
 	then
-		echo -e "\t${cyan}Found and editing file: ${CYAN}$1${NC}"
+		echo -e "  ${cyan}Found and editing file: ${CYAN}$1${NC}"
 		echo -e "\n#OXISCRIPTS HEADER (remove only as block!)" >> $1
 		echo "if [ -f $TARGETDIR/init.sh ]; then" >> $1
 		echo "       [ -z \"\$PS1\" ] && return" >> $1
 		echo "       . $TARGETDIR/init.sh" >> $1
 		echo "fi" >> $1
 	else
-		echo -e "\t${cyan}Found but not editing file: ${CYAN}$1${NC}"
+		echo -e "  ${cyan}Found but not editing file: ${CYAN}$1${NC}"
 	fi
 }
 
@@ -241,15 +253,30 @@ for FILE in $(ls /home/*/.bash_history); do
 done
 
 
-echo -e "\n${cyan}Checking for optional apps (These are only needed if you plan to use some of the modules/functions):\n\t${RED}\c"
+install=""
+doit="0"
+echo -e "\n${cyan}Checking optional apps (These are only needed if you plan to use some of the modules/functions):\n  ${RED}\c"
 BINS="rdiff-backup fdupes rsync mailx screen"
 for BIN in $BINS
 do
 	if [ ! -n "$(which $BIN 2>/dev/null )" ]; then
 		echo -e "$BIN \c"
+		install="$install$BIN "
+		doit="1"
 	fi
 done
 echo -e "${NC}"
+
+if [ "$doit" == "1" ];
+then
+	if [ "$LSBID" == "debian" ];
+	then
+		aptitude install $install -P
+	elif [ "$LSBID" == "gentoo" ];
+	then
+		emerge $install -av
+	fi
+fi
 
 echo -e "\n${BLUE}Everything done.${NC}\n\n${RED}Please configure your jobs in $TARGETDIR/jobs!${NC}\n"
 . /etc/oxiscripts/init.sh
